feat(progress): add a cron job to handle calculating progress

+++ src/main/java/com/swpteam/smokingcessation/domain/entity/Member.java
@@ -29,8 +29,8 @@ public class Member extends AuditableEntity {
     int score;
     int highestStreak;
     LocalDateTime lastCounterReset;
-    float cigarettesAvoided;
-    float moneySaved;
+    double cigarettesAvoided;
+    double moneySaved;
 
     @Enumerated(EnumType.STRING)
     MemberGender gender;
